refactoring Player & Phantom controllers
1. CharacterBase라는 하나의 higher class를 만듬 -> state들도 새로 만듬
2. PhantomController refactoring (TurnManager에서 관리하도록 함)
~> 아직 버그는 수정하지 못함
