+++ Chronus/Assets/Prefabs/Phantom.prefab
@@ -1152,6 +1152,10 @@ GameObject:
   - component: {fileID: 3266294757220750305}
   - component: {fileID: 7332264089347800436}
   - component: {fileID: 9059034893468094471}
+  - component: {fileID: -4016232556565193883}
+  - component: {fileID: -2377306329990543250}
+  - component: {fileID: -9102312364794136100}
+  - component: {fileID: 3486113093813852857}
   m_Layer: 3
   m_Name: Phantom
   m_TagString: Player
@@ -1243,16 +1247,68 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   animator: {fileID: 0}
-  curHopDir: 1
-  curTurnAngle: 0
   playerCurPos: {x: 0, y: 0, z: 0}
   playerCurRot: {x: 0, y: 0, z: 0, w: 0}
-  listCommandOrder: []
   moveSpeedHor: 6
   moveSpeedVer: 3
   turnSpeed: 480
+  curHopDir: 1
+  curTurnAngle: 0
+  cIdle: {fileID: -4016232556565193883}
+  cMove: {fileID: -2377306329990543250}
+  cTurn: {fileID: -9102312364794136100}
+  cHop: {fileID: 3486113093813852857}
   doneAction: 0
-  phantomStates: {fileID: 0}
+  listCommandOrder: []
+  order: 0
+--- !u!114 &-4016232556565193883
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4137369883225489573}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 310756e53545e7345b0027b8c50b0c0f, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &-2377306329990543250
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4137369883225489573}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f2543d0a56684f946a8ae90bd38438a0, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &-9102312364794136100
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4137369883225489573}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: abe693de95a814942a04ea235b0b704b, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &3486113093813852857
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4137369883225489573}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 3d6abbc146afa994e8bc7da6d06b66d6, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &4313637727183743810
 GameObject:
   m_ObjectHideFlags: 0
