+++ Chronus/Assets/Prefabs/Player.prefab
@@ -2118,6 +2118,10 @@ GameObject:
   - component: {fileID: 7338755304510778482}
   - component: {fileID: 6194105668740652634}
   - component: {fileID: 52053991078801657}
+  - component: {fileID: 1256036149688309480}
+  - component: {fileID: -520197802082359419}
+  - component: {fileID: 756250703899310066}
+  - component: {fileID: 8506127019799237812}
   - component: {fileID: 8813005280719247447}
   m_Layer: 3
   m_Name: Player
@@ -2210,17 +2214,69 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   animator: {fileID: 0}
-  curHopDir: 0
-  curTurnAngle: 0
   playerCurPos: {x: 0, y: 0, z: 0}
   playerCurRot: {x: 0, y: 0, z: 0, w: 0}
-  listCommandLog: []
   moveSpeedHor: 6
   moveSpeedVer: 3
   turnSpeed: 480
+  curHopDir: 0
+  curTurnAngle: 0
+  cIdle: {fileID: 1256036149688309480}
+  cMove: {fileID: -520197802082359419}
+  cTurn: {fileID: 756250703899310066}
+  cHop: {fileID: 8506127019799237812}
   doneAction: 0
-  playerStates: {fileID: 0}
-  listCommandOrderForPhantom: []
+  isPhantomExists: 0
+  isTimeRewinding: 0
+  listCommandLog: []
+--- !u!114 &1256036149688309480
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5664397843930143027}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 310756e53545e7345b0027b8c50b0c0f, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &-520197802082359419
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5664397843930143027}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f2543d0a56684f946a8ae90bd38438a0, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &756250703899310066
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5664397843930143027}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: abe693de95a814942a04ea235b0b704b, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &8506127019799237812
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5664397843930143027}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 3d6abbc146afa994e8bc7da6d06b66d6, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!114 &8813005280719247447
 MonoBehaviour:
   m_ObjectHideFlags: 0
