+++ Chronus/Assets/Scripts/Character.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3268ff631bc0a694387dab4c502f39fa
+guid: c81e79477cd687c41b5dec656a9f5b66
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
