+++ Chronus/Assets/Scripts/Character/CharacterHop.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 06abf6e752cc5dd4698ca7daac3b3f7f
+guid: 3d6abbc146afa994e8bc7da6d06b66d6
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
