+++ Chronus/Assets/Scripts/Character/CharacterIdle.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 510d68839e4c8b04f87b3a1de4de9f01
+guid: 310756e53545e7345b0027b8c50b0c0f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
