+++ Chronus/Assets/Scripts/Character/CharacterMove.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6ae2bcb157a81f643b58bdde02838081
+guid: f2543d0a56684f946a8ae90bd38438a0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
