+++ Chronus/Assets/Scripts/CharacterBase.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3c74e4496e10aeb4db414076140b0565
+guid: 0d85e3381cf93414180149e3627cfb21
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
