--- Chronus/Assets/Scripts/Phantom.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f91ee4bdfb2f3874eac6f4ad376764ed
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
