--- Chronus/Assets/Scripts/Phantom/PhantomTurn.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 77ca11a03c40f2141a1c490d6136fb47
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
