--- Chronus/Assets/Scripts/Player/PlayerMove.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 01fefa60fd7e52b4585d23c9f465d675
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
