--- Chronus/Assets/Scripts/Player/PlayerTurn.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 788b943b9ad67cc44a3a321753aa4ba6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
